+++ microgradx/nn/__init__.py
@@ -1,7 +1,7 @@
 from microgradx.nn.module import Module, Sequential, ModuleList
 from microgradx.nn.linear import Linear
 from microgradx.nn.conv import Conv2d, MaxPool2d, Flatten
-from microgradx.nn.norm import LayerNorm, RMSNorm
+from microgradx.nn.norm import LayerNorm, RMSNorm, BatchNorm1d, BatchNorm2d
 from microgradx.nn.dropout import Dropout
 from microgradx.nn.activation import ReLU, GELU, Sigmoid, Tanh, Softmax
 from microgradx.nn.embedding import Embedding
@@ -14,7 +14,7 @@ from microgradx.nn import init
 __all__ = [
     "Module", "Sequential", "ModuleList",
     "Linear", "Conv2d", "MaxPool2d", "Flatten",
-    "LayerNorm", "RMSNorm",
+    "LayerNorm", "RMSNorm", "BatchNorm1d", "BatchNorm2d",
     "Dropout",
     "ReLU", "GELU", "Sigmoid", "Tanh", "Softmax",
     "Embedding",